chore: update stacks-rpc-client

+++ components/chainhook-sdk/src/indexer/stacks/mod.rs
@@ -332,7 +332,7 @@ pub fn standardize_stacks_block(
         .into();
     let current_len = u64::saturating_sub(
         block.burn_block_height,
-        1 + chain_ctx.pox_info.first_burnchain_block_height,
+        1 + (chain_ctx.pox_info.first_burnchain_block_height as u64),
     );
     let pox_cycle_id: u32 = (current_len / pox_cycle_length).try_into().unwrap_or(0);
     let mut events: HashMap<&String, Vec<&NewEvent>> = HashMap::new();
